# doc/technical/openapi.yaml
openapi: 3.1.0
info:
  title: 'SoManAgent API'
  version: 1.0.0
  description: 'Versioned OpenAPI contract for the SoManAgent backend. This specification is spec-first: operations may be documented ahead of implementation when they explicitly set `x-somanagent-implemented` to `false`.'
servers:
  - url: 'http://localhost:8080'
...
paths:
...
  /api/roles:
    get:
      tags:
        - Roles
      summary: 'List roles'
      operationId: listRoles
      responses:
        200:
          $ref: '#/components/responses/Ok'
    post:
      tags:
        - Roles
      summary: 'Create a role'
      operationId: createRole
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        201:
          $ref: '#/components/responses/Created'
  '/api/roles/{id}':
    get:
      tags:
        - Roles
      summary: 'Get a role'
      operationId: getRole
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
    patch:
      tags:
        - Roles
      summary: 'Update a role'
      operationId: updateRole
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
    delete:
      tags:
        - Roles
      summary: 'Delete a role'
      operationId: deleteRole
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
  '/api/roles/{id}/skills':
    post:
      tags:
        - Roles
      summary: 'Attach a skill to a role'
      operationId: addRoleSkill
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
  '/api/roles/{id}/skills/{skillId}':
    delete:
      tags:
        - Roles
      summary: 'Detach a skill from a role'
      operationId: removeRoleSkill
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/SkillIdPath'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
  /api/skills:
    get:
      tags:
        - Skills
      summary: 'List skills'
      operationId: listSkills
      responses:
        200:
          $ref: '#/components/responses/Ok'
    post:
      tags:
        - Skills
      summary: 'Create a skill'
      operationId: createSkill
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        201:
          $ref: '#/components/responses/Created'
  '/api/skills/{id}':
    get:
      tags:
        - Skills
      summary: 'Get a skill'
      operationId: getSkill
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
    delete:
      tags:
        - Skills
      summary: 'Delete a skill'
      operationId: deleteSkill
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
  /api/skills/import:
    post:
      tags:
        - Skills
      summary: 'Import a skill'
      operationId: importSkill
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        201:
          $ref: '#/components/responses/Created'
  '/api/skills/{id}/content':
    patch:
      tags:
        - Skills
      summary: 'Update skill content'
      operationId: updateSkillContent
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SkillIdPath:
      name: skillId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
    ArbitraryJsonBody:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArbitraryJsonObject'
...
  responses:
    Ok:
      description: 'Successful response'
    Created:
      description: 'Resource created'
...
    NoContent:
      description: 'Resource deleted successfully'
  schemas:
...
    ArbitraryJsonObject:
      type: object
      additionalProperties: true
...